Convert ServiceCHelper into CDI beans
Demoing this in talks with the ServiceCHelper can be hard for the
audience because when you show serviceCHelper there are lots of
methods and it is hard to read the right bit of code. Instead if
it is broken up into multiple beans you can show only the code that
is required. Then to switch between the different demos you can do
this by using a different injection qualifier.